refactor: extract login and welcome UI
